+++ packages/react-components/src/components/utils/SendBoxUtils.ts
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+const EMPTY_MESSAGE_REGEX = /^\s*$/;
+const MAXIMUM_LENGTH_OF_MESSAGE = 8000;
+
+const sanitizeText = (message: string): string => {
+  if (EMPTY_MESSAGE_REGEX.test(message)) {
+    return '';
+  } else {
+    return message;
+  }
+};
